backend/requirements.txt
# Core
django==5.0.4
psycopg2-binary==2.9.9
gunicorn==22.0.0

# API
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Tarefas Assíncronas (Setup inicial)
celery==5.4.0
redis==5.0.4
django-redis==5.4.0

# Utils
python-dotenv==1.0.1